+++ api-gateway/src/main/java/com/example/api/gateway/ApiGatewayApplication.java
@@ -3,9 +3,11 @@ package com.example.api.gateway;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.cloud.netflix.hystrix.EnableHystrix;
 
 @SpringBootApplication
-@EnableEurekaClient//Even a api gateway is also a client to service registry
+@EnableEurekaClient//Even an api gateway is also a client to service registry
+@EnableHystrix
 public class ApiGatewayApplication {
 
 	public static void main(String[] args) {