+++ api-gateway/src/main/java/com/example/api/gateway/config/MainConfig.java
@@ -1,0 +1,18 @@
+package com.example.api.gateway.config;
+
+import com.netflix.hystrix.HystrixCommand;
+import com.netflix.hystrix.HystrixCommandGroupKey;
+import com.netflix.hystrix.HystrixCommandProperties;
+import org.springframework.cloud.client.circuitbreaker.Customizer;
+import org.springframework.cloud.netflix.hystrix.HystrixCircuitBreakerFactory;
+import org.springframework.context.annotation.Bean;
+
+public class MainConfig {
+//    @Bean
+//    public Customizer<HystrixCircuitBreakerFactory> defaultConfig() {
+//        return factory -> factory.configureDefault(id -> HystrixCommand.Setter
+//                .withGroupKey(HystrixCommandGroupKey.Factory.asKey(id))
+//                .andCommandPropertiesDefaults(HystrixCommandProperties.Setter()
+//                        .withExecutionTimeoutInMilliseconds(4000)));
+//    }
+}